sailfish: hammerhead: tidyup resolutions.ini
Rule of thumb: Pick 2-3 aspect ratios and have 2-3 resolutions per
aspect ratio. As for videos, you can put up to 4 resolutions.

If a non-standard aspect ratio (not 16:9 or 4:3) has only 1 resolution,
do not pick it, especially if it's a low-res one, and you have a handful
of other aspect ratios for that mode already.

Signed-off-by: Simonas Leleiva <simonas.leleiva@jollamobile.com>